[WOOR-63] refactor: DTO 패키지 이동
